move component and store. move osmImporter to microservice and add a store to it
